% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.sbtab.R
\name{merge.sbtab}
\alias{merge.sbtab}
\title{Merge multiple SBtab files into seperate dataframes per table type}
\usage{
\method{merge}{sbtab}(filelist, outputname, outputdir)
}
\arguments{
\item{filelist}{A list of filepaths to TSV formatted SBtabs to merge}

\item{outputname}{A string which each merged file begins with}

\item{outputdir}{The output folder for the merged file}
}
\value{
.rds file inside output directory
}
\description{
Merge multiple SBtab files into seperate dataframes per table type
}
\examples{
merge.sbtab(c("data-raw/physmap6.tsv", "data-raw/physmap7.tsv"), "merged_physmap67", "data/merged")

merge.sbtab(c("data-raw/physmap6.tsv",
"data-raw/physmap7.tsv",
"data-raw/physmap8.tsv",
"data-raw/physmap9.tsv"), "merged_physmap6789", "data/merged")
}
