% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_prepped.R
\name{merge_prepped}
\alias{merge_prepped}
\title{Merge prepped SBtab dataframes made with sep_sbtab() and prep_merge()}
\usage{
merge_prepped(fls, oname, odir = getwd())
}
\arguments{
\item{fls}{A list of prepped files to merge (all have to be the same type! i.e. reactions, edges, etc.)}

\item{oname}{The outputname of the merged file}

\item{odir}{The outputfolder}
}
\value{
A .rds file containing the merged dataframes from that type
}
\description{
Merge prepped SBtab dataframes made with sep_sbtab() and prep_merge()
}
\examples{
prepped_physmap6_reactions <- system.file("extdata",
"prepped_physmap6_reactions.rds", package="sbtabr")
prepped_physmap7_reactions <- system.file("extdata",
"prepped_physmap7_reactions.rds", package="sbtabr")
merge_prepped(c(prepped_physmap6_reactions, prepped_physmap7_reactions),
'merged_physmap67_reactions')
}
