% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_net_graph.R
\name{merged_net_graph}
\alias{merged_net_graph}
\title{Visualize the merged files made with merge_sbtab}
\usage{
merged_net_graph(edge_file, species_file, compartments_file)
}
\arguments{
\item{edge_file}{the output file from the merge_sbtab function or an existing edge file}

\item{species_file}{the output file from the merge_sbtab function or an existing species file}

\item{compartments_file}{the output file from the merge_sbtab function or an existing compartments file}
}
\value{
A network graph for merged files
}
\description{
Visualize the merged files made with merge_sbtab
}
\examples{
merged_net_graph(merged_physmap67_edges, merged_physmap67_species, merged_physmap67_compartments)
}
