% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{merged_physmap6789_compartments}
\alias{merged_physmap6789_compartments}
\title{The merged compartments tables from SBtab files physmap6, 7, 8 and 9}
\description{
This is a .rds file containing all the compartments data of four SBtab files
The original SBtab files are physmap6, physmap7, physmap8 and physmap9
The file has been created using the merge_sbtab() function
}
\section{merged_physmap6789_compartments.rds}{


This data is used in merged_net_graph().
}

