% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{merged_physmap67_edges}
\alias{merged_physmap67_edges}
\title{The merged edges tables from SBtab files physmap6 and physmap7}
\description{
This is a .rds file containing all the edges data of two SBtab files
The original SBtab files are physmap6 and physmap7
The file has been created using the merge_sbtab() function
}
\section{merged_physmap67_edges.rds}{


This data is used in merged_net_graph().
}

