% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_graph.R
\name{net_graph}
\alias{net_graph}
\title{Title}
\usage{
net_graph(edge_list, species_list, compartments_list)
}
\arguments{
\item{edge_list}{the output file from the sep_sbtab function or an existing edge file}

\item{species_list}{the output file from the sep_sbtab function or an existing species file}

\item{compartments_list}{the output file from the sep_sbtab function or an existing compartments file}
}
\value{
A network graph for unmerged files
}
\description{
Title
}
\examples{
netgraph("data/edges/physmap6_edges.rds", "data/species/physmap6_species.rds", "data/compartments/physmap6_compartments.rds")

netgraph("data/edges/physmap7_edges.rds", "data/species/physmap7_species.rds", "data/compartments/physmap7_compartments.rds")
}
