% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{physmap6}
\alias{physmap6}
\title{Original SBtab file 'physmap6'}
\description{
This is a .tsv file containing the original physmap6 SBtab
It came from the SysRev platform and was provided to this project by M. Teunis
}
\section{physmap6.tsv}{


This data is used in sep_sbtab() and merge_sbtab().
}

