% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap6_species}
\alias{physmap6_species}
\title{physmap6_species}
\format{
A file path to a .rds file containing a dataframe
The data frame has 11 rows and 8 columns:
\describe{
\item{Name}{Name of the compound}
\item{ID}{Species ID}
\item{Type}{Type of species, i.e. protein or gene}
\item{Location}{Location ID as given in compartments file}
\item{IsConstant}{Logical value indicating constant status}
\item{InitialConcentration}{Optional value of concentration}
\item{HasOnlySubstanceUnits}{Logical value indicating substance units}
\item{Resolver}{Database ID's}
...
}
}
\source{
<data-raw/physmap6.tsv>
}
\usage{
physmap6_species
}
\description{
Result of putting the physmap6.tsv file through sep_sbtab
}
\keyword{datasets}
