% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{physmap7}
\alias{physmap7}
\title{Original SBtab file 'physmap7'}
\description{
This is a .tsv file containing the original physmap7 SBtab
It came from the SysRev platform and was provided to this project by M. Teunis
}
\section{physmap7.tsv}{


This data is used in sep_sbtab() and merge_sbtab().
}

