% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap9}
\alias{physmap9}
\title{physmap9}
\format{
A file path to a .tsv file containing an SBtab
}
\source{
<data-raw/physmap9.tsv>
}
\usage{
physmap9
}
\description{
An original SBtab file as provided by the SysRev platform
}
\keyword{datasets}
