% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.merge.R
\name{prep.merge}
\alias{prep.merge}
\title{Prepare SBtab dataframes made with sep.sbtab() for merging}
\usage{
prep.merge(
  fname,
  ogid,
  odir = getwd(),
  reactionsdata = FALSE,
  edgesdata = FALSE,
  speciesdata = FALSE
)
}
\arguments{
\item{fname}{The file path of the file to be prepped}

\item{ogid}{The identifier or name of the original file the file came from}

\item{odir}{The outputdirectory}

\item{reactionsdata}{Asks if the file contains reactions (i.e. multiple columns containing IDs)}

\item{edgesdata}{Asks if the file contains edges (i.e. multiple columns containing IDs)}

\item{speciesdata}{Asks if the file contains species (i.e. multiple columns containing IDs)}
}
\value{
.rds files prepared for merging
}
\description{
Prepare SBtab dataframes made with sep.sbtab() for merging
}
\examples{
prep.merge("data/compartments/physmap8_compartments.rds",
"physmap8", "data/compartments")
prep.merge("data/edges/physmap8_edges.rds", "physmap8",
"data/edges", edgesdata = TRUE)
prep.merge("data/reactions/physmap8_reactions.rds", "physmap8",
"data/reactions", reactionsdata = TRUE)
prep.merge("data/species/physmap8_species.rds", "physmap8",
"data/species", speciesdata = TRUE)
}
