% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_merge.R
\name{prep_merge}
\alias{prep_merge}
\title{Prepare SBtab dataframes made with sep_sbtab for merging}
\usage{
prep_merge(
  fname,
  ogid,
  odir = getwd(),
  reactionsdata = FALSE,
  edgesdata = FALSE,
  speciesdata = FALSE
)
}
\arguments{
\item{fname}{The file path of the file to be prepped}

\item{ogid}{The identifier or name of the original file the file came from}

\item{odir}{The outputdirectory}

\item{reactionsdata}{Asks if the file contains reactions (i.e. multiple columns containing IDs)}

\item{edgesdata}{Asks if the file contains edges (i.e. multiple columns containing IDs)}

\item{speciesdata}{Asks if the file contains species (i.e. multiple columns containing IDs)}
}
\value{
.rds files prepared for merging
}
\description{
Prepare SBtab dataframes made with sep_sbtab for merging
}
\examples{
physmap6_species <- system.file("extdata", "physmap6_species.rds", package="PackageONTOX")
prep_merge(physmap6_species, 'physmap6', speciesdata = TRUE)
}
