% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{prepped_physmap6_reactions}
\alias{prepped_physmap6_reactions}
\title{The reactions table from SBtab file physmap6 prepared for merging}
\description{
This is a .rds file containing the species of SBtab file physmap6
The file has been prepared for merging using prep_merge()
The dataframe has an extra column containing the original filename
The ID's have been made unique by adding the original filename
}
\section{prepped_physmap6_reactions.rds}{


This data is used in merge_prepped().
}

