% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prepped_physmap7_reactions}
\alias{prepped_physmap7_reactions}
\title{prepped_physmap7_reactions}
\format{
A file path to a .rds file containing a dataframe
The data frame has ... rows and ... columns:
\describe{
\item{Compounds}{Compound ID's}
\item{Products}{Product ID's}
\item{ID}{Reaction ID}
\item{Compoundname}{Compound name}
\item{Productname}{Product name}
\item{Productname}{Name of original file}
...
}
}
\source{
<data-raw/physmap7.tsv>
}
\usage{
prepped_physmap7_reactions
}
\description{
Result of putting the physmap7.tsv file through sep_sbtab and prep_merge
}
\keyword{datasets}
