% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_sbtab.R
\name{sep.sbtab}
\alias{sep.sbtab}
\title{Title}
\usage{
sep.sbtab(fname, oname, odir = getwd(), colnum = 25)
}
\arguments{
\item{fname}{A tsv file path}

\item{oname}{The outputname}

\item{odir}{The output directory}

\item{colnum}{The number of columns of the widest dataframe}
}
\value{
Seperates SBtab files into reactions, species and compounds, also saves a dataframe with full infoormation.
}
\description{
Title
}
\examples{
sep.sbtab("SBtab_examples/physmap7.tsv", "physmap7", "data")

sep.sbtab("SBtab_examples/physmap10.tsv", "physmap10", "data")
}
