% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_sbtab.R
\name{sep_sbtab}
\alias{sep_sbtab}
\title{Seperate SBtab files into seperate dataframes per table}
\usage{
sep_sbtab(fname, oname, odir = getwd(), colnum = 25)
}
\arguments{
\item{fname}{A tsv file path}

\item{oname}{The outputname}

\item{odir}{The output directory}

\item{colnum}{The number of columns of the widest dataframe}
}
\value{
Seperates SBtab files into reactions, species and compounds, also saves a dataframe with full infoormation.
}
\description{
Seperate SBtab files into seperate dataframes per table
}
\examples{
physmap6 <- system.file("extdata", "physmap6.tsv", package="PackageONTOX")
physmap7 <- system.file("extdata", "physmap7.tsv", package="PackageONTOX")
sep_sbtab(physmap6, "physmap6")
sep_sbtab(physmap7, "physmap7")
}
