% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysrev_to_tsv_xml.R
\name{sysrev_to_tsv_xml}
\alias{sysrev_to_tsv_xml}
\title{Converts sysrev file into SBtab and SBML file}
\usage{
sysrev_to_tsv_xml(pid, aid, token, documentname)
}
\arguments{
\item{pid}{Identifier of used papers}

\item{aid}{Identifier of paper authors}

\item{token}{Token}

\item{documentname}{Output document name}
}
\value{
Two .rds files, SBtab and SBML
}
\description{
Converts sysrev file into SBtab and SBML file
}
\examples{
To do
}
