% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.prepped.R
\name{merge.prepped}
\alias{merge.prepped}
\title{Merge prepped SBtab dataframes made with sep.sbtab() and prep.merge()}
\usage{
\method{merge}{prepped}(fls, oname, odir = getwd())
}
\arguments{
\item{fls}{A list of prepped files to merge (all have to be the same type! i.e. reactions, edges, etc.)}

\item{oname}{The outputname of the merged file}

\item{odir}{The outputfolder}
}
\value{
A .rds file containing the merged dataframes
}
\description{
Merge prepped SBtab dataframes made with sep.sbtab() and prep.merge()
}
\examples{
merge.prepped(edgeslist, "phys67_edges", "data/merged")

merge.prepped(reactionslist, "phys6789_reactions", "data/merged")
}
