% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_prepped.R
\name{merge.prepped}
\alias{merge.prepped}
\title{Merge prepped SBtab dataframes made with SBtab_to_dfs() and premerge()}
\usage{
\method{merge}{prepped}(fls, oname, odir = getwd())
}
\arguments{
\item{fls}{A list of prepped files to merge (all have to be the same type!)}

\item{oname}{The outputname of the merged file}

\item{odir}{The (already existing) outputfolder}
}
\value{
A .rds file
}
\description{
Merge prepped SBtab dataframes made with SBtab_to_dfs() and premerge()
}
\examples{
merge.prepped(edgeslist, "phys67_edges", "data/merged")

merge.prepped(reactionslist, "phys6789_reactions", "data/merged")
}
