% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sbtab.R
\name{merge.sbtab}
\alias{merge.sbtab}
\title{merge.sbtab}
\usage{
\method{merge}{sbtab}(filelist, outputname, outputdir)
}
\arguments{
\item{filelist}{A list of filepaths to TSV formatted SBtabs to merge}

\item{outputname}{A string which each merged file begins with}

\item{outputdir}{The output folder for the merged file}
}
\value{
.rds file inside output directory
}
\description{
merge.sbtab
}
\examples{
merge.sbtab(c("SBtab_examples/physmap6.tsv", "SBtab_examples/physmap7.tsv"), "merged67", "data/merged")

merge.sbtab(c("SBtab_examples/physmap6.tsv",
"SBtab_examples/physmap7.tsv",
"SBtab_examples/physmap8.tsv",
"SBtab_examples/physmap9.tsv",
"SBtab_examples/physmap10.tsv"), "merged678910", "data/merged")
}
