% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_prepped.R
\name{merge_prepped}
\alias{merge_prepped}
\title{Merge prepped SBtab dataframes made with sep.sbtab() and prep.merge()}
\usage{
merge_prepped(fls, oname, odir = getwd())
}
\arguments{
\item{fls}{A list of prepped files to merge (all have to be the same type! i.e. reactions, edges, etc.)}

\item{oname}{The outputname of the merged file}

\item{odir}{The outputfolder}
}
\value{
A .rds file containing the merged dataframes from that type
}
\description{
Merge prepped SBtab dataframes made with sep.sbtab() and prep.merge()
}
\examples{
merge_prepped(c("data/edges/prepped_physmap6_edges.rds", "data/edges/prepped_physmap7_edges.rds"),
"merged_physmap67_edges", "data/merged")
merge_prepped(c("data/species/prepped_physmap6_species.rds", "data/species/prepped_physmap7_species.rds"),
"merged_physmap67_species", "data/merged")
}
