% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sbtab.R
\name{merge_sbtab}
\alias{merge_sbtab}
\title{Merge multiple SBtab files into seperate dataframes per table type}
\usage{
merge_sbtab(filelist, outputname, outputdir = getwd())
}
\arguments{
\item{filelist}{A list of filepaths to TSV formatted SBtabs to merge}

\item{outputname}{A string which each merged file begins with}

\item{outputdir}{The output folder for the merged file}
}
\value{
.rds file inside output directory
}
\description{
Merge multiple SBtab files into seperate dataframes per table type
}
\examples{
merge_sbtab(c(physmap6, physmap7, physmap8, physmap9), 'merged_physmap6789')
}
