% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_net_graph.R
\name{merged_net_graph}
\alias{merged_net_graph}
\title{Title}
\usage{
merged_net_graph(edge_list, species_list, compartments_list)
}
\arguments{
\item{edge_list}{the output file from the merge_sbtab function or an existing edge file}

\item{species_list}{the output file from the merge_sbtab function or an existing species file}

\item{compartments_list}{the output file from the merge_sbtab function or an existing compartments file}
}
\value{
A network graph for merged files
}
\description{
Title
}
\examples{
merged_net_graph("data/merged/merged_physmap67_edge_list.rds", "data/merged/merged_physmap67_species.rds", "data/merged/merged_physmap67_compartments.rds")

merged_net_graph("data/merged/merged_physmap789_edge_list.rds", "data/merged/merged_physmap789_species.rds", "data/merged/merged_physmap789_compartments.rds")
}
