% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_physmap6789_compartments}
\alias{merged_physmap6789_compartments}
\title{merged_physmap6789_compartments}
\format{
A file path to a .rds file containing a dataframe
The data frame has ... rows and ... columns:
\describe{
\item{ID}{Given compartment ID}
\item{Name}{Compartment name}
\item{Size}{Relative size of compartment}
\item{Unit}{Unit the size is based on}
\item{filename}{Name of original file}
...
}
}
\source{
<data-raw/physmap7.tsv>
}
\usage{
merged_physmap6789_compartments
}
\description{
Result of putting physmap6, 7, 8 and 9 through merge_sbtab
}
\keyword{datasets}
