% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_physmap6789_compartments}
\alias{merged_physmap6789_compartments}
\title{merged_physmap6789_compartments}
\format{
\subsection{\code{merged_physmap6789_compartments}}{

A data frame with 14 rows and 6 columns:
\describe{
\item{Compounds}{Compound ID's}
\item{Products}{Product ID's}
\item{ID}{Reaction ID}
\item{Compoundname}{Compound name}
\item{Productname}{Product name}
\item{Productname}{Name of original file}
...
}
}
}
\source{
<data-raw/physmap7.tsv>
}
\usage{
merged_physmap6789_compartments
}
\description{
Result of putting physmap7 through sep.sbtab and prep.merge
}
\keyword{datasets}
