% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_physmap6789_edges}
\alias{merged_physmap6789_edges}
\title{merged_physmap6789_edges}
\format{
A file path to a .rds file containing a dataframe
The data frame has ... rows and ... columns:
\describe{
\item{Compounds}{Compound ID's}
\item{Products}{Product ID's}
\item{ID}{Reaction ID}
\item{Compoundname}{Compound name}
\item{Productname}{Product name}
\item{filename}{Name of original file}
...
}
}
\source{
<data-raw/physmap7.tsv>
}
\usage{
merged_physmap6789_edges
}
\description{
Result of putting physmap6, 7, 8 and 9 through merge_sbtab
}
\keyword{datasets}
