% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_physmap6789_species}
\alias{merged_physmap6789_species}
\title{merged_physmap6789_species}
\format{
A file path to a .rds file containing a dataframe
The data frame has ... rows and ... columns:
\describe{
\item{Name}{Name of the compound}
\item{ID}{Species ID}
\item{Type}{Type of species, i.e. protein or gene}
\item{Location}{Location ID as given in compartments file}
\item{IsConstant}{Logical value indicating constant status}
\item{InitialConcentration}{Optional value of concentration}
\item{HasOnlySubstanceUnits}{Logical value indicating substance units}
\item{Resolver}{Database ID's}
\item{filename}{Name of original file}
\item{InitialValue}{States active form}
...
}
}
\source{
<data-raw/physmap7.tsv>
}
\usage{
merged_physmap6789_species
}
\description{
Result of putting physmap6, 7, 8 and 9 through merge_sbtab
}
\keyword{datasets}
