% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{merged_physmap67_compartments}
\alias{merged_physmap67_compartments}
\title{The merged compartments tables from SBtab files physmap6 and physmap7}
\description{
This is a .rds file containing all the compartments data of two SBtab files
The original SBtab files are physmap6 and physmap7
The file has been created using the merge_sbtab() function
}
\section{merged_physmap67_compartments.rds}{


This data is used in merged_net_graph().
}

