% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_graph.R
\name{net_graph}
\alias{net_graph}
\title{net_graph}
\usage{
net_graph(edge_file, species_file, compartments_file)
}
\arguments{
\item{edge_file}{the output file from the sep_sbtab function or an existing edge file}

\item{species_file}{the output file from the sep_sbtab function or an existing species file}

\item{compartments_file}{the output file from the sep_sbtab function or an existing compartments file}
}
\value{
A network graph for unmerged files
}
\description{
net_graph
}
