% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_graph.R
\name{net_graph}
\alias{net_graph}
\title{Visualize a SBtab file made with sep_sbtab}
\usage{
net_graph(edge_file, species_file, compartments_file)
}
\arguments{
\item{edge_file}{the output file from the sep_sbtab function or an existing edge file}

\item{species_file}{the output file from the sep_sbtab function or an existing species file}

\item{compartments_file}{the output file from the sep_sbtab function or an existing compartments file}
}
\value{
A network graph for unmerged files
}
\description{
Visualize a SBtab file made with sep_sbtab
}
\examples{
physmap6_edges <- system.file("extdata", "physmap6_edges.rds", package="PackageONTOX")
physmap6_species <- system.file("extdata", "physmap6_species.rds", package="PackageONTOX")
physmap6_compartments <- system.file("extdata", "physmap6_compartments.rds", package="PackageONTOX")
net_graph(physmap6_edges, physmap6_species, physmap6_compartments)
}
