% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_graph.R
\name{net_graph}
\alias{net_graph}
\title{Title}
\usage{
net_graph(edge_list, species_list)
}
\arguments{
\item{edge_list}{A list for the lines between nodes}

\item{species_list}{A list with the nodes}
}
\value{
A graph with the nodes species_list connected by lines from edge_list
}
\description{
Title
}
\examples{
net_graph(physmap8_edges, physmap8_species)

net_graph(physmap9_edges, physmap9_species)
}
