% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap6}
\alias{physmap6}
\title{physmap6}
\format{
\subsection{\code{physmap6}}{

A data frame with 14 rows and 6 columns:
\describe{
\item{Compounds}{Compound ID's}
...
}
}
}
\source{
<data-raw/physmap6.tsv>
}
\usage{
physmap6
}
\description{
An original SBtab file as provided by the SysRev platform
}
\keyword{datasets}
