% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap6_compartments}
\alias{physmap6_compartments}
\title{physmap6_compartments}
\format{
A file path to a .rds file containing a dataframe
The data frame has 2 rows and 4 columns:
\describe{
\item{ID}{Given compartment ID}
\item{Name}{Compartment name}
\item{Size}{Relative size of compartment}
\item{Unit}{Unit the size is based on}
...
}
}
\source{
<data-raw/physmap6.tsv>
}
\usage{
physmap6_compartments
}
\description{
Result of putting the physmap6.tsv file through sep_sbtab
}
\keyword{datasets}
