% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap6_edges}
\alias{physmap6_edges}
\title{physmap6_edges}
\format{
\subsection{\code{physmap6_edges}}{

A data frame with 14 rows and 6 columns:
\describe{
\item{Compounds}{Compound ID's}
\item{Products}{Product ID's}
\item{ID}{Reaction ID}
\item{Compoundname}{Compound name}
\item{Productname}{Product name}
\item{Productname}{Name of original file}
...
}
}
}
\source{
<data-raw/physmap6.tsv>
}
\usage{
physmap6_edges
}
\description{
Result of putting physmap6 through sep.sbtab and prep.merge
}
\keyword{datasets}
