% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap6_species}
\alias{physmap6_species}
\title{physmap6_species}
\format{
A file path to a .rds file containing a dataframe
The data frame has ... rows and ... columns:
\describe{
\item{Compounds}{Compound ID's}
\item{Products}{Product ID's}
\item{ID}{Reaction ID}
\item{Compoundname}{Compound name}
\item{Productname}{Product name}
\item{Productname}{Name of original file}
...
}
}
\source{
<data-raw/physmap6.tsv>
}
\usage{
physmap6_species
}
\description{
Result of putting the physmap6.tsv file through sep_sbtab
}
\keyword{datasets}
