% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{physmap6_species}
\alias{physmap6_species}
\title{The species table from SBtab file physmap6}
\description{
This is a .rds file containing the species of SBtab file physmap6
The file has been created using the sep_sbtab() function
}
\section{physmap6_species.rds}{


This data is used in net_graph().
}

