% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap7}
\alias{physmap7}
\title{physmap7}
\format{
A file path to a .tsv file containing an SBtab
}
\source{
<data-raw/physmap7.tsv>
}
\usage{
physmap7
}
\description{
An original SBtab file as provided by the SysRev platform
}
\keyword{datasets}
