% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{physmap8}
\alias{physmap8}
\title{Original SBtab file 'physmap8'}
\description{
This is a .tsv file containing the original physmap8 SBtab
It came from the SysRev platform and was provided to this project by M. Teunis
}
\section{physmap8.tsv}{


This data is used in sep_sbtab() and merge_sbtab().
}

