% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap8}
\alias{physmap8}
\title{physmap8}
\format{
\subsection{\code{physmap8}}{

A data frame with 14 rows and 6 columns:
\describe{
\item{Compounds}{Compound ID's}
...
}
}
}
\source{
<data-raw/physmap8.tsv>
}
\usage{
physmap8
}
\description{
An original SBtab file as provided by the SysRev platform
}
\keyword{datasets}
