% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physmap9}
\alias{physmap9}
\title{physmap9}
\format{
\subsection{\code{physmap9}}{

A data frame with 14 rows and 6 columns:
\describe{
\item{Compounds}{Compound ID's}
...
}
}
}
\source{
<data-raw/physmap9.tsv>
}
\usage{
physmap9
}
\description{
An original SBtab file as provided by the SysRev platform
}
\keyword{datasets}
