% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{physmap9}
\alias{physmap9}
\title{Original SBtab file 'physmap9'}
\description{
This is a .tsv file containing the original physmap9 SBtab
It came from the SysRev platform and was provided to this project by M. Teunis
}
\section{physmap9.tsv}{


This data is used in sep_sbtab() and merge_sbtab().
}

