% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prepped_physmap7_reactions}
\alias{prepped_physmap7_reactions}
\title{prepped_physmap7_reactions}
\format{
A file path to a .rds file containing a dataframe
The data frame has 8 rows and 7 columns:
\describe{
\item{ID}{Reaction ID}
\item{Type}{Type of reaction}
\item{IsReversible}{Logical value to indicate reversibility}
\item{Reactants}{Reactant ID's}
\item{Products}{Product ID's}
\item{Modifiers}{Modifier ID's}
\item{filename}{Name of original file}
...
}
}
\source{
<data-raw/physmap7.tsv>
}
\usage{
prepped_physmap7_reactions
}
\description{
Result of putting the physmap7.tsv file through sep_sbtab and prep_merge
}
\keyword{datasets}
