% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prepped_physmap7_reactions}
\alias{prepped_physmap7_reactions}
\title{prepped_physmap7_reactions}
\format{
\subsection{\code{prepped_physmap7_reactions}}{

A data frame with 14 rows and 6 columns:
\describe{
\item{Compounds}{Compound ID's}
\item{Products}{Product ID's}
\item{ID}{Reaction ID}
\item{Compoundname}{Compound name}
\item{Productname}{Product name}
\item{Productname}{Name of original file}
...
}
}
}
\source{
<data-raw/physmap7.tsv>
}
\usage{
prepped_physmap7_reactions
}
\description{
Result of putting physmap7 through sep.sbtab and prep.merge
}
\keyword{datasets}
